fix(snuggle): Sauron audit fixes — precision, rotation wiring, OOB logging
Triple coordinate audit findings addressed:
- mm_to_scaled() now takes double (was float, lost precision from double intermediates)
- snuggle_lock_rotation properly wired (was using only allow_rotations)
- OOB log message uses rotation-aware variables (was referencing removed vars)

Sauron triple audit result: 10/10 coordinate boundaries correct,
Z offset confirmed non-issue (ArrangePolygon.translation is 2D by design,
Z preserved on ModelInstance by apply_arrange_result not touching it).

+++ src/libslic3r/Arrange/SnuggleTransform.cpp
@@ -30,12 +30,12 @@ float scaled_to_mm(coord_t val)
 }
 
 // ── Snuggle mm -> Orca scaled coords ────────────────────────────────
-// Input space:  Snuggle mm (float).
+// Input space:  Snuggle mm (double for precision on large beds).
 // Output space: Orca scaled integer (coord_t). 1 unit = SCALING_FACTOR mm.
-coord_t mm_to_scaled(float val_mm)
+coord_t mm_to_scaled(double val_mm)
 {
     // scale: val_mm / SCALING_FACTOR -> scaled integer
-    coord_t result = scaled<coord_t>(static_cast<double>(val_mm));
+    coord_t result = scaled<coord_t>(val_mm);
 
     BOOST_LOG_TRIVIAL(trace) << "[SnuggleXform] mm_to_scaled: "
         << val_mm << " mm -> " << result << " scaled"
